new Entites (PlanElementOpeningHoursEntity, PlanElementTypeEntity)

+++ src/TripMaker.Core/ExternalServices.Entities/GooglePlaceDetails/GooglePlaceDetailsResult.cs
@@ -19,7 +19,7 @@ namespace TripMaker.ExternalServices.Entities.GooglePlaceDetails
         public List<Photo> photos { get; set; }
         public string place_id { get; set; }
         public double? rating { get; set; }
-        public double? price_level { get; set; }
+        public int? price_level { get; set; }
         public string reference { get; set; }
         public List<Review> reviews { get; set; }
         public string scope { get; set; }
